Clean up teacher grading decision endpoints
