+++ services/main-frontend/src/services/backend/answers-requiring-attention.ts
@@ -1,14 +1,7 @@
 import { mainFrontendClient } from "../mainFrontendClient"
 
-import {
-  AnswersRequiringAttention,
-  NewTeacherGradingDecision,
-  UserExerciseState,
-} from "@/shared-module/common/bindings"
-import {
-  isAnswersRequiringAttention,
-  isUserExerciseState,
-} from "@/shared-module/common/bindings.guard"
+import { AnswersRequiringAttention } from "@/shared-module/common/bindings"
+import { isAnswersRequiringAttention } from "@/shared-module/common/bindings.guard"
 import { validateResponse } from "@/shared-module/common/utils/fetching"
 
 export const fetchAnswersRequiringAttention = async (
@@ -23,18 +16,4 @@ export const fetchAnswersRequiringAttention = async (
     },
   )
   return validateResponse(response, isAnswersRequiringAttention)
-}
-
-export const updateAnswerRequiringAttention = async ({
-  user_exercise_state_id,
-  exercise_id,
-  action,
-  manual_points,
-  hidden,
-}: NewTeacherGradingDecision): Promise<UserExerciseState> => {
-  const response = await mainFrontendClient.put(
-    `/exercise-slide-submissions/update-answer-requiring-attention`,
-    { user_exercise_state_id, exercise_id, action, manual_points, hidden },
-  )
-  return validateResponse(response, isUserExerciseState)
 }
